requirements.txt
# Requirements for Multi-Agent VLN Navigation System

# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Transformers and models
transformers>=4.30.0
tokenizers>=0.13.0

# Habitat simulation (optional - install separately)
# habitat-sim>=0.3.0
# habitat-lab>=0.3.0

# Computer vision
opencv-python>=4.8.0
pillow>=10.0.0

# APIs
openai>=1.0.0
anthropic>=0.5.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Logging and monitoring
colorlog>=6.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pandas>=2.0.0
h5py>=3.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional)
# mypy>=1.4.0

# Documentation (optional)
# sphinx>=7.0.0
# sphinx-rtd-theme>=1.3.0

# Development tools (optional)
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0